NAT穿透

+++ client/chart.c
@@ -244,7 +244,7 @@ static gint voice_button_release_event(GtkWidget *widget, GdkEventButton *event,
             char hexKey[65] = {0};
             for (int i = 0; i < 32; ++i)
             {
-                sprintf(hexKey + i * 2, "%02x", entry->key[i]);
+                sprintf(hexKey + i * 2, "%02x",(int) entry->key[i]);
             }
             log_info("Discover", "Try to register key %s\n", hexKey);
